+++ app/endpoints/categories.py
@@ -6,7 +6,7 @@ from flask_restplus import Resource
 from app.models import db, Category
 from app.serializers import category
 from app.restplus import API
-from .auth import authorization_required
+from app.helpers import authorization_required
 
 # Linting exceptions
